+++ src/assets/fonts/lulo-clean/lulo-clean-one-bold.otf
